[MOVE/#52] explore 폴더 이동
