+++ app/src/main/java/com/napzak/market/presentation/explore/explore/component/GenreSearchNoticeSection.kt
@@ -1,4 +1,4 @@
-package com.napzak.market.presentation.explore.component
+package com.napzak.market.presentation.explore.explore.component
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
