+++ app/src/main/java/com/napzak/market/presentation/explore/explore/state/ExploreUiState.kt
@@ -1,4 +1,4 @@
-package com.napzak.market.presentation.explore.state
+package com.napzak.market.presentation.explore.explore.state
 
 import androidx.compose.runtime.Immutable
 import com.napzak.market.core.common.state.UiState
@@ -6,7 +6,7 @@ import com.napzak.market.core.type.SortType
 import com.napzak.market.core.type.TradeType
 import com.napzak.market.domain.explore.model.ProductItem
 import com.napzak.market.domain.genre.model.Genre
-import com.napzak.market.presentation.explore.type.ExploreScreenType
+import com.napzak.market.presentation.explore.explore.type.ExploreScreenType
 
 @Immutable
 data class ExploreUiState(
