+++ app/src/main/java/com/napzak/market/presentation/home/state/HomeUiState.kt
@@ -3,7 +3,7 @@ package com.napzak.market.presentation.home.state
 import com.napzak.market.R
 import com.napzak.market.core.common.state.UiState
 import com.napzak.market.domain.explore.model.ProductItem
-import com.napzak.market.presentation.explore.type.TradeType
+import com.napzak.market.presentation.explore.explore.type.TradeType
 
 data class HomeUiState(
     val bannerImages: UiState<List<Int>> = UiState.Loading,
