+++ app/src/main/java/com/napzak/market/presentation/main/type/MainTab.kt
@@ -7,7 +7,7 @@ import com.napzak.market.R
 import com.napzak.market.core.common.navigation.MainTabRoute
 import com.napzak.market.core.common.navigation.Route
 import com.napzak.market.presentation.dummy.navigation.Dummy
-import com.napzak.market.presentation.explore.navigation.Explore
+import com.napzak.market.presentation.explore.explore.navigation.Explore
 import com.napzak.market.presentation.home.navigation.Home
 import com.napzak.market.presentation.mypage.navigation.MyPage
