+++ app/src/main/java/com/napzak/market/presentation/marketinfo/component/MarketInfoBottomSheetScreen.kt
@@ -8,8 +8,8 @@ import androidx.compose.ui.window.PopupProperties
 import com.napzak.market.core.type.BottomSheetType
 import com.napzak.market.core.type.SortType
 import com.napzak.market.domain.genre.model.Genre
-import com.napzak.market.presentation.explore.bottomSheet.GenreSearchBottomSheet
-import com.napzak.market.presentation.explore.bottomSheet.SortBottomSheet
+import com.napzak.market.presentation.explore.explore.bottomSheet.GenreSearchBottomSheet
+import com.napzak.market.presentation.explore.explore.bottomSheet.SortBottomSheet
 import com.napzak.market.presentation.marketinfo.state.MarketInfoBottomSheetState
 
 @Composable
